+++ src/OuterList.php
@@ -62,18 +62,15 @@ final class OuterList implements MemberList
      */
     public static function fromHttpValue(Stringable|string $httpValue): self
     {
-        return new self(...(function (iterable $members) {
-            foreach ($members as $member) {
-                if (!is_array($member[0])) {
-                    yield Item::fromAssociative(...$member);
+        $converter = fn (array $member): InnerList|Item => match (true) {
+            is_array($member[0]) => InnerList::fromAssociative(
+                array_map(fn (array $item) => Item::fromAssociative(...$item), $member[0]),
+                $member[1]
+            ),
+            default => Item::fromAssociative($member[0], $member[1]),
+        };
 
-                    continue;
-                }
-
-                $member[0] = array_map(fn (array $item) => Item::fromAssociative(...$item), $member[0]);
-                yield InnerList::fromAssociative(...$member);
-            }
-        })(Parser::parseList($httpValue)));
+        return new self(...array_map($converter, Parser::parseList($httpValue)));
     }
 
     /**
@@ -124,8 +121,9 @@ final class OuterList implements MemberList
 
     public function has(string|int ...$keys): bool
     {
+        $max = count($this->members);
         foreach ($keys as $offset) {
-            if (null === $this->filterIndex($offset)) {
+            if (null === $this->filterIndex($offset, $max)) {
                 return false;
             }
         }
@@ -133,13 +131,13 @@ final class OuterList implements MemberList
         return [] !== $keys;
     }
 
-    private function filterIndex(string|int $index): int|null
+    private function filterIndex(string|int $index, int|null $max = null): int|null
     {
         if (!is_int($index)) {
             return null;
         }
 
-        $max = count($this->members);
+        $max ??= count($this->members);
 
         return match (true) {
             [] === $this->members,
@@ -155,12 +153,7 @@ final class OuterList implements MemberList
      */
     public function get(string|int $key): StructuredField
     {
-        $index = $this->filterIndex($key);
-        if (null === $index) {
-            throw InvalidOffset::dueToIndexNotFound($key);
-        }
-
-        return $this->members[$index];
+        return $this->members[$this->filterIndex($key) ?? throw InvalidOffset::dueToIndexNotFound($key)];
     }
 
     /**
@@ -170,11 +163,10 @@ final class OuterList implements MemberList
      */
     public function unshift(iterable|StructuredField|Token|ByteSequence|DateTimeInterface|string|int|float|bool ...$members): static
     {
-        if ([] === $members) {
-            return $this;
-        }
-
-        return new self(...array_values($members), ...$this->members);
+        return match (true) {
+            [] === $members => $this,
+            default => new self(...array_values($members), ...$this->members),
+        };
     }
 
     /**
@@ -184,11 +176,10 @@ final class OuterList implements MemberList
      */
     public function push(iterable|StructuredField|Token|ByteSequence|DateTimeInterface|string|int|float|bool ...$members): static
     {
-        if ([] === $members) {
-            return $this;
-        }
-
-        return new self(...$this->members, ...array_values($members));
+        return match (true) {
+            [] === $members => $this,
+            default => new self(...$this->members, ...array_values($members)),
+        };
     }
 
     /**
@@ -200,10 +191,9 @@ final class OuterList implements MemberList
      */
     public function insert(int $key, iterable|StructuredField|Token|ByteSequence|DateTimeInterface|string|int|float|bool ...$members): static
     {
-        $offset = $this->filterIndex($key);
+        $offset = $this->filterIndex($key) ?? throw InvalidOffset::dueToIndexNotFound($key);
 
         return match (true) {
-            null === $offset => throw InvalidOffset::dueToIndexNotFound($key),
             0 === $offset => $this->unshift(...$members),
             count($this->members) === $offset => $this->push(...$members),
             [] === $members => $this,
@@ -220,37 +210,35 @@ final class OuterList implements MemberList
      */
     public function replace(int $key, iterable|StructuredField|Token|ByteSequence|DateTimeInterface|string|int|float|bool $member): static
     {
-        if (null === ($offset = $this->filterIndex($key))) {
-            throw InvalidOffset::dueToIndexNotFound($key);
-        }
-
+        $offset = $this->filterIndex($key) ?? throw InvalidOffset::dueToIndexNotFound($key);
         $member = self::filterMember($member);
-        if ($member->toHttpValue() === $this->members[$offset]->toHttpValue()) {
-            return $this;
-        }
 
-        return new self(...array_replace($this->members, [$offset => $member]));
+        return match (true) {
+            $member->toHttpValue() === $this->members[$offset]->toHttpValue() => $this,
+            default => new self(...array_replace($this->members, [$offset => $member])),
+        };
     }
 
     public function remove(string|int ...$keys): static
     {
+        $max = count($this->members);
         $offsets = array_filter(
             array_map(
-                fn (int $index): int|null => $this->filterIndex($index),
+                fn (int $index): int|null => $this->filterIndex($index, $max),
                 array_filter($keys, static fn (string|int $key): bool => is_int($key))
             ),
             fn (int|null $index): bool => null !== $index
         );
 
-        if ([] === $offsets) {
-            return $this;
-        }
-
-        return new self(...array_filter(
-            $this->members,
-            fn (int $key): bool => !in_array($key, $offsets, true),
-            ARRAY_FILTER_USE_KEY
-        ));
+        return match (true) {
+            [] === $offsets => $this,
+            $max === count($offsets) => new self(),
+            default => new self(...array_filter(
+                $this->members,
+                fn (int $key): bool => !in_array($key, $offsets, true),
+                ARRAY_FILTER_USE_KEY
+            )),
+        };
     }
 
     /**