+++ src/Parameters.php
@@ -13,6 +13,7 @@ use function array_keys;
 use function array_map;
 use function count;
 use function implode;
+use function is_int;
 use function is_string;
 
 /**
@@ -49,7 +50,7 @@ final class Parameters implements MemberOrderedMap
     private static function filterMember(mixed $member): object
     {
         return match (true) {
-            $member instanceof ValueAccess && $member instanceof ParameterAccess => $member->parameters()->hasNoMembers() ? $member : throw new InvalidArgument('The "'.$member::class.'" instance is not a Bare Item.'),
+            $member instanceof ParameterAccess && $member instanceof ValueAccess => $member->parameters()->hasNoMembers() ? $member : throw new InvalidArgument('The "'.$member::class.'" instance is not a Bare Item.'),
             $member instanceof StructuredField => throw new InvalidArgument('An instance of "'.$member::class.'" can not be a member of "'.self::class.'".'),
             default => Item::new($member),
         };
@@ -87,15 +88,14 @@ final class Parameters implements MemberOrderedMap
      */
     public static function fromPairs(iterable $pairs): self
     {
-        if ($pairs instanceof MemberOrderedMap) {
-            return new self($pairs);
-        }
-
-        return new self((function (iterable $pairs) {
-            foreach ($pairs as [$key, $member]) {
-                yield $key => $member;
-            }
-        })($pairs));
+        return match (true) {
+            $pairs instanceof MemberOrderedMap => new self($pairs),
+            default => new self((function (iterable $pairs) {
+                foreach ($pairs as [$key, $member]) {
+                    yield $key => $member;
+                }
+            })($pairs)),
+        };
     }
 
     /**
@@ -146,7 +146,7 @@ final class Parameters implements MemberOrderedMap
     }
 
     /**
-     * @return Iterator<array{0:string, 1:SfItem}>
+     * @return Iterator<int, array{0:string, 1:SfItem}>
      */
     public function toPairs(): Iterator
     {
@@ -175,24 +175,20 @@ final class Parameters implements MemberOrderedMap
     }
 
     /**
-     * @throws SyntaxError   If the key is invalid
      * @throws InvalidOffset If the key is not found
      *
      * @return SfItem
      */
     public function get(string|int $key): StructuredField
     {
-        if (!$this->has($key)) {
-            throw InvalidOffset::dueToKeyNotFound($key);
-        }
-
-        return $this->members[$key];
+        return $this->members[$key] ?? throw InvalidOffset::dueToKeyNotFound($key);
     }
 
     public function hasPair(int ...$indexes): bool
     {
+        $max = count($this->members);
         foreach ($indexes as $index) {
-            if (null === $this->filterIndex($index)) {
+            if (null === $this->filterIndex($index, $max)) {
                 return false;
             }
         }
@@ -203,28 +199,27 @@ final class Parameters implements MemberOrderedMap
     /**
      * Filters and format instance index.
      */
-    private function filterIndex(int $index): int|null
+    private function filterIndex(int $index, int|null $max = null): int|null
     {
-        $max = count($this->members);
+        $max ??= count($this->members);
 
         return match (true) {
-            [] === $this->members, 0 > $max + $index, 0 > $max - $index - 1 => null,
+            [] === $this->members,
+            0 > $max + $index,
+            0 > $max - $index - 1 => null,
             0 > $index => $max + $index,
             default => $index,
         };
     }
 
     /**
-     * @throws InvalidOffset if the index is not found
+     * @throws InvalidOffset If the key is not found
      *
      * @return array{0:string, 1:SfItem}
      */
     public function pair(int $index): array
     {
-        $offset = $this->filterIndex($index);
-        if (null === $offset) {
-            throw InvalidOffset::dueToIndexNotFound($index);
-        }
+        $offset = $this->filterIndex($index) ?? throw InvalidOffset::dueToIndexNotFound($index);
 
         return [...$this->toPairs()][$offset];
     }
@@ -242,47 +237,55 @@ final class Parameters implements MemberOrderedMap
      */
     private function newInstance(array $members): self
     {
-        if ($members == $this->members) {
-            return $this;
-        }
-
-        return new self($members);
+        return match(true) {
+            $members == $this->members => $this,
+            default => new self($members),
+        };
     }
 
     public function remove(string|int ...$keys): static
     {
-        /** @var array<array-key, true> $indexes */
-        $indexes = array_fill_keys($keys, true);
-        $pairs = [];
-        foreach ($this->toPairs() as $index => $pair) {
-            if (!isset($indexes[$index]) && !isset($indexes[$pair[0]])) {
-                $pairs[] = $pair;
-            }
-        }
-
-        if (count($this->members) === count($pairs)) {
+        if ([] === $this->members || [] === $keys) {
             return $this;
         }
 
-        return self::fromPairs($pairs);
+        $offsets = array_keys($this->members);
+        $max = count($offsets);
+        $reducer = fn (array $carry, string|int $key): array => match (true) {
+            is_string($key) && (false !== ($position = array_search($key, $offsets, true))),
+            is_int($key) && (null !== ($position = $this->filterIndex($key, $max))) => [$position => true] + $carry,
+            default => $carry,
+        };
+
+        $indices = array_reduce($keys, $reducer, []);
+
+        return match (true) {
+            [] === $indices => $this,
+            $max === count($indices) => self::new(),
+            default => self::fromPairs((function (array $offsets) {
+                foreach ($this->toPairs() as $offset => $pair) {
+                    if (!array_key_exists($offset, $offsets)) {
+                        yield $pair;
+                    }
+                }
+            })($indices)),
+        };
     }
 
     public function append(string $key, StructuredField|Token|ByteSequence|DateTimeInterface|string|int|float|bool $member): static
     {
         $members = $this->members;
         unset($members[$key]);
-        $members[MapKey::from($key)->value] = self::filterMember($member);
 
-        return $this->newInstance($members);
+        return $this->newInstance([...$members, MapKey::from($key)->value => self::filterMember($member)]);
     }
 
     public function prepend(string $key, StructuredField|Token|ByteSequence|DateTimeInterface|string|int|float|bool $member): static
     {
         $members = $this->members;
         unset($members[$key]);
-        $members = [MapKey::from($key)->value => self::filterMember($member), ...$members];
 
-        return $this->newInstance($members);
+        return $this->newInstance([MapKey::from($key)->value => self::filterMember($member), ...$members]);
     }
 
     /**
@@ -290,16 +293,13 @@ final class Parameters implements MemberOrderedMap
      */
     public function push(array ...$pairs): self
     {
-        if ([] === $pairs) {
-            return $this;
-        }
-
-        $newPairs = iterator_to_array($this->toPairs());
-        foreach ($pairs as $pair) {
-            $newPairs[] = $pair;
-        }
-
-        return self::fromPairs($newPairs);
+        return match (true) {
+            [] === $pairs => $this,
+            default => self::fromPairs((function (iterable $pairs) {
+                yield from $this->toPairs();
+                yield from $pairs;
+            })($pairs)),
+        };
     }
 
     /**
@@ -307,15 +307,13 @@ final class Parameters implements MemberOrderedMap
      */
     public function unshift(array ...$pairs): self
     {
-        if ([] === $pairs) {
-            return $this;
-        }
-
-        foreach ($this->members as $key => $member) {
-            $pairs[] = [$key, $member];
-        }
-
-        return self::fromPairs($pairs);
+        return match (true) {
+            [] === $pairs => $this,
+            default => self::fromPairs((function (iterable $pairs) {
+                yield from $pairs;
+                yield from $this->toPairs();
+            })($pairs)),
+        };
     }
 
     /**
@@ -323,10 +321,9 @@ final class Parameters implements MemberOrderedMap
      */
     public function insert(int $index, array ...$members): static
     {
-        $offset = $this->filterIndex($index);
+        $offset = $this->filterIndex($index) ?? throw InvalidOffset::dueToIndexNotFound($index);
 
         return match (true) {
-            null === $offset => throw InvalidOffset::dueToIndexNotFound($index),
             [] === $members => $this,
             0 === $offset => $this->unshift(...$members),
             count($this->members) === $offset => $this->push(...$members),
@@ -344,18 +341,14 @@ final class Parameters implements MemberOrderedMap
      */
     public function replace(int $index, array $member): static
     {
-        $offset = $this->filterIndex($index);
-        if (null === $offset) {
-            throw InvalidOffset::dueToIndexNotFound($index);
-        }
-
+        $offset = $this->filterIndex($index) ?? throw InvalidOffset::dueToIndexNotFound($index);
         $member[1] = self::filterMember($member[1]);
         $pairs = iterator_to_array($this->toPairs());
-        if ($pairs[$offset] == $member) {
-            return $this;
-        }
 
-        return self::fromPairs(array_replace($pairs, [$offset => $member]));
+        return match (true) {
+            $pairs[$offset] == $member => $this,
+            default => self::fromPairs(array_replace($pairs, [$offset => $member])),
+        };
     }
 
     /**